requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.6.0
chardet==4.0.0
click==8.1.3
discord.py==1.7.3
feedparser==6.0.10
flake8==5.0.4
idna==3.3
mccabe==0.7.0
more-itertools==8.13.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.20.0
PyYAML==6.0
sgmllib3k==1.0.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.8.1
